Extract route definitions into a separate file
