+++ aiohttp/web.py
@@ -7,7 +7,7 @@ from importlib import import_module
 
 from . import (helpers, web_app, web_exceptions, web_fileresponse,
                web_middlewares, web_protocol, web_request, web_response,
-               web_runner, web_server, web_urldispatcher, web_ws)
+               web_routedef, web_runner, web_server, web_urldispatcher, web_ws)
 from .log import access_logger
 from .web_app import *  # noqa
 from .web_exceptions import *  # noqa
@@ -16,6 +16,7 @@ from .web_middlewares import *  # noqa
 from .web_protocol import *  # noqa
 from .web_request import *  # noqa
 from .web_response import *  # noqa
+from .web_routedef import *  # noqa
 from .web_runner import *  # noqa
 from .web_runner import AppRunner, GracefulExit, SockSite, TCPSite, UnixSite
 from .web_server import *  # noqa
@@ -28,6 +29,7 @@ __all__ = (web_protocol.__all__ +
            web_fileresponse.__all__ +
            web_request.__all__ +
            web_response.__all__ +
+           web_routedef.__all__ +
            web_exceptions.__all__ +
            web_urldispatcher.__all__ +
            web_ws.__all__ +
